move verify_no_duplicate_relationships
